fix: Migrate to impermanence home manager module v2
After some trial and error I found that the root cause of the issues
I've been having with sleep on both of my laptops is that the current
Impermanence home manager module uses fileSystems instead of
systemd.mounts (and systemd doesn't like the fileSystems mounts for some
reason). There's a lot of additional benefits to this migration too,
like being able to actually hide the mounts so they don't show up in my
files program.

I also removed the explicit SuspendState=mem as I believe Linux will
automatically choose deep sleep if available, and I think a lot of the
issues I was having was as a result of me trying to force suspend states
rather than just letting the OS choose.

One issue which I didn't predict as a result of this change is that now
my keyboard layout is different in the initial greeter login. I'll fix
that later.